Refactor brand controller
